remove unnecessary files and rename custom button
