--- code/frontend/src/components/buttonPrimary.css
@@ -1,26 +1,0 @@
-.login-button {
-    display: inline-flex;
-    padding: auto;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-    padding-left: 1rem;
-    padding-right: 1rem;
-    width: fit-content;
-
-    margin: 1rem;
-    border-radius: 10px;
-    border: 2px solid var(--Rotkehlchen-orange-default);
-    background: var(--Rotkehlchen-yellow-default);
-  }
-  
-  .login-button:hover {
-    border-radius: 10px;
-    border: 2px solid var(--Rotkehlchen-orange-default);
-    background: var(--Rotkehlchen-yellow-hover);
-  }
-  .login-button:active {
-    border-radius: 10px;
-    border: 2px solid var(--Rotkehlchen-orange-default);
-    background: var(--Rotkehlchen-orange-default);
-  }
