+++ code/frontend/src/styles/colors.css
@@ -1,5 +1,6 @@
 :root {
   --Rotkehlchen-gray: #e7ecf2;
+  --Rotkehlchen-gray-hover: #D5D7DA;;
   --Rotkehlchen-brown-light: #a28d7a;
   --Rotkehlchen-brown-middle: #7f6d6d;
   --Rotkehlchen-brown-dark: #4c4141;
